Moved HIE specific commands to a new module, cleaned up haddock(no initialization required anymore)
