+++ src/Haskell/Ide/Engine/Plugin/Haddock.hs
@@ -3,14 +3,12 @@
 {-# LANGUAGE RankNTypes          #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE FlexibleContexts    #-}
-{-# LANGUAGE CPP                 #-}
 module Haskell.Ide.Engine.Plugin.Haddock where
 
 import           Control.Monad.State
 import           Data.Foldable
 import qualified Data.Map                                     as Map
 import           Data.Monoid
-import           Data.Maybe
 import qualified Data.Text                                    as T
 import           Data.IORef
 import           System.Directory
@@ -18,16 +16,13 @@ import           System.FilePath
 import           GHC
 import           GhcMonad
 import qualified GhcMod.Monad                                 as GM
+import qualified GhcMod.LightGhc                              as GM
 import           Haskell.Ide.Engine.MonadTypes
-import           Haskell.Ide.Engine.MonadFunctions            as MF
 import           HscTypes
 import           Name
 import           Packages
-#if __GLASGOW_HASKELL__ > 800
-import           NameCache
-#endif
 
-import           Haskell.Ide.Engine.Plugin.HaRe
+import           Haskell.Ide.Engine.Plugin.HieExtras
 
 import Documentation.Haddock
 import Documentation.Haddock.Types
@@ -69,20 +64,6 @@ lookupSrcHtmlForModule df m = do
     mn = map (\x -> if x == '.' then '-' else x) mns
     mns = moduleNameString $ moduleName m
 
-newtype HaddockState = HS { nameCache :: Maybe (IORef NameCache)}
-instance ExtensionClass HaddockState where
-  initialValue = HS Nothing
-
-initialized :: MonadMTState IdeState m => m Bool
-initialized = do
-  HS x <- MF.get
-  return (isJust x)
-
-initializeHaddock :: IdeGhcM ()
-initializeHaddock = do
-  ref <- GM.unGmlT $ withSession (return . hsc_NC)
-  MF.put $ HS $ Just ref
-
 nameCacheFromGhcMonad :: GhcMonad m => NameCacheAccessor m
 nameCacheFromGhcMonad = ( read_from_session , write_to_session )
   where
@@ -93,26 +74,28 @@ nameCacheFromGhcMonad = ( read_from_session , write_to_session )
        ref <- withSession (return . hsc_NC)
        liftIO $ writeIORef ref nc'
 
+runInLightGhc :: GM.LightGhc a -> IdeM a
+runInLightGhc a = do
+  hscEnvRef <- ghcSession <$> readMTS
+  mhscEnv <- liftIO $ traverse readIORef hscEnvRef
+  case mhscEnv of
+    Nothing -> error "Ghc Session not initialized"
+    Just env -> GM.runLightGhc env a
+
 nameCacheFromIdeM :: NameCacheAccessor IdeM
 nameCacheFromIdeM = ( read_from_session , write_to_session )
   where
-    read_from_session = do
-       HS (Just ref) <- MF.get
-       liftIO $ readIORef ref
-    write_to_session nc' = do
-       HS (Just ref) <- MF.get
-       liftIO $ writeIORef ref nc'
+    read_from_session = runInLightGhc $ fst nameCacheFromGhcMonad
+    write_to_session _ = error "tried to write to namecache(only reads allowed)"
 
 
 getDocsForName :: DynFlags -> Name -> IdeM (Maybe T.Text)
 getDocsForName df name = do
-  init' <- initialized
   let mfs = nameModule_maybe name >>=
               lookupHaddock df . moduleUnitId
   mf <- case mfs of
     Nothing -> pure Nothing
-    Just fs | not init' -> pure Nothing
-            | otherwise -> liftIO $ do
+    Just fs -> liftIO $ do
       fs' <- filterM doesFileExist fs
       case fs' of
         [] -> pure Nothing
